requirements.txt
flask
flask_wtf
flask_login
sqlalchemy
orm
werkzeug
datetime
werkzeug
base64
PIL
io
wtforms
os
shutil
pymorphy
sqlalchemy_serializer